Move cleanableDestinationListFromMounts to run_common.go
